Adiciona aggregates

+++ bank-account/account.cmd/src/main/java/com/techbank/account/cmd/api/commands/OpenAccountCommand.java
@@ -4,10 +4,12 @@ import com.techbank.account.common.dto.AccountType;
 import com.techbank.cqrs.core.commands.BaseCommand;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OpenAccountCommand extends BaseCommand {
 
     private String accountHolder;
     private AccountType accountType;
-    private double openingBalance;
+    private BigDecimal openingBalance;
 }
